Removed all warning 1.044 2/7/2023

+++ BiquOperatorPanel/Megasonic/TouchLCD/GUI/PanelMain.event.h
@@ -4,17 +4,17 @@
 extern uint8_t* MemoryDumpDisplayAddress;
 void initializeSoapNudList();
 
-void glBtnPlus_TouchEvent(void* sender, uint16_t x, uint16_t y);
-void glBtnMinus_TouchEvent(void* sender, uint16_t x, uint16_t y);
-void glBtnSave_TouchEvent(void* sender, uint16_t x, uint16_t y);
-void glBtnApply_TouchEvent(void* sender, uint16_t x, uint16_t y);
+void glBtnPlus_TouchEvent(void* sender, long unsigned int x, long unsigned int y);
+void glBtnMinus_TouchEvent(void* sender, long unsigned int x, long unsigned int y);
+void glBtnSave_TouchEvent(void* sender, long unsigned int x, long unsigned int y);
+void glBtnApply_TouchEvent(void* sender, long unsigned int x, long unsigned int y);
 
-void eraseFlashButonEvent(void* sender, uint16_t x, uint16_t y);
-void displaySoapstringButtonEvent(void* sender, uint16_t x, uint16_t y);
-void ShiftbitWritetestButtonEvent(void* sender, uint16_t x, uint16_t y);
-void displayCalibrationButtonEvent(void* sender, uint16_t x, uint16_t y);
-void LoadSoapstringButtonEvent(void* sender, uint16_t x, uint16_t y);
-void SaveSoapstringButtonEvent(void* sender, uint16_t x, uint16_t y);
+void eraseFlashButonEvent(void* sender, long unsigned int x, long unsigned int y);
+void displaySoapstringButtonEvent(void* sender, long unsigned int x, long unsigned int y);
+void ShiftbitWritetestButtonEvent(void* sender, long unsigned int x, long unsigned int y);
+void displayCalibrationButtonEvent(void* sender, long unsigned int x, long unsigned int y);
+void LoadSoapstringButtonEvent(void* sender, long unsigned int x, long unsigned int y);
+void SaveSoapstringButtonEvent(void* sender, long unsigned int x, long unsigned int y);
 
 void ChangeDisplayModeListEvent(void* sender);
